Move funcs to top level import
