+++ tests/unit/synthesizers/test_generate.py
@@ -2,10 +2,10 @@ from unittest.mock import Mock
 
 import pytest
 
+from sdgym import create_sdv_synthesizer_variant, create_single_table_synthesizer
 from sdgym.synthesizers import FastMLPreset, SDVRelationalSynthesizer, SDVTabularSynthesizer
 from sdgym.synthesizers.generate import (
-    SYNTHESIZER_MAPPING, create_multi_table_synthesizer, create_sdv_synthesizer_variant,
-    create_sequential_synthesizer, create_single_table_synthesizer)
+    SYNTHESIZER_MAPPING, create_multi_table_synthesizer, create_sequential_synthesizer)
 
 
 def test_create_single_table_synthesizer():
