+++ tests/unit/test_datasets.py
@@ -6,9 +6,8 @@ from zipfile import ZipFile
 import botocore
 import pandas as pd
 
-from sdgym.datasets import (
-    _get_bucket_name, _get_dataset_path, download_dataset, get_available_datasets,
-    get_dataset_paths)
+from sdgym import get_available_datasets
+from sdgym.datasets import _get_bucket_name, _get_dataset_path, download_dataset, get_dataset_paths
 
 
 class AnyConfigWith:
